requirements.txt
pyautogui==0.9.54
pyperclip==1.8.2
openai==0.27.0
python-dotenv==0.19.2